requirements.txt
customtkinter
tkcalendar
babel
python-dotenv